move caching logic to agentstore
